+++ src/js/handler/modal/onSearchClip.js
@@ -1,20 +1,20 @@
-import { getQueryString } from '../utils/getQueryString.js';
+import { getQueryString } from '../../utils/getQueryString.js';
 import {
   YOUTUBE,
   MESSAGE,
   LOCAL_STORAGE_KEY,
   BASE_URL,
-} from '../utils/constant.js';
-import { request } from '../utils/fetch.js';
-import { showElement, hideElement } from '../utils/setAttribute.js';
-import storage from '../utils/localStorage.js';
+} from '../../utils/constant.js';
+import { request } from '../../utils/fetch.js';
+import { showElement, hideElement } from '../../utils/setAttribute.js';
+import storage from '../../utils/localStorage.js';
 import {
   clearSearchResult,
   renderClips,
   renderRecentKeywords,
-} from '../view/modal.js';
-import { snackbar } from '../utils/snackbar.js';
-import $DOM from '../utils/DOM.js';
+} from '../../view/modal.js';
+import { snackbar } from '../../utils/snackbar.js';
+import $DOM from '../../utils/DOM.js';
 
 const getRecentKeywords = (keyword) => {
   const keywords = storage.get(LOCAL_STORAGE_KEY.RECENT_KETWORDS) ?? [];
